update file
function that searches

for a value in an array

of integers using

the Linear search algorithm

+++ 0x1E-search_algorithms/0-linear.c
@@ -10,7 +10,7 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	int i;
+	size_t i;
 
 	if (array == NULL)
 	{
@@ -18,9 +18,9 @@ int linear_search(int *array, size_t size, int value)
 	}
 	if (array)
 	{
-		for (i = 0; i < (int)size ; i++)
+		for (i = 0; i < size ; i++)
 		{
-			printf("Value checked array [%d] = [%d]\n", i, array[i]);
+			printf("Value checked array [%lu] = [%d]\n", i, (int)array[i]);
 			if (array[i] == value)
 			{
 				return ((int)i);
